Moved code around
Large refactor part, may be broken
